+++ src/RumiApplication.php
@@ -48,9 +48,6 @@ class RumiApplication extends Application
         $this->setUpCommands();
     }
 
-    /**
-     * @return ContainerBuilder
-     */
     private function initContainer()
     {
         $this->container = new ContainerBuilder();
@@ -65,9 +62,14 @@ class RumiApplication extends Application
 
     private function setUpCommands()
     {
-        $oRunCommand = new RunCommand($this->container, $this->container->get('trivago.rumi.event_dispatcher'));
+        $oRunCommand = new RunCommand(
+            $this->container->get('trivago.rumi.event_dispatcher'),
+            $this->container->get('trivago.rumi.services.config_reader'),
+            $this->container->get('trivago.rumi.commands.run.stage_executor')
+        );
+
         $this->add($oRunCommand);
-        $this->add(new CheckoutCommand($this->container->get('trivago.rumi.process.git_processes_execution')));
+        $this->add(new CheckoutCommand($this->container->get('trivago.rumi.process.git_processes_execution'), $this->container->get('trivago.rumi.process.git_clone_process')));
         $this->add(new CacheStoreCommand($this->container));
         $this->add(new CacheRestoreCommand($this->container));
         $this->setDefaultCommand($oRunCommand->getName());