+++ tests/Models/CacheConfigTest.php
@@ -16,25 +16,25 @@
  * limitations under the License.
  */
 
-namespace Trivago\Rumi\Commands;
+namespace Trivago\Rumi\Models;
 
-
-use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputOption;
-
-abstract class CommandAbstract extends Command
+/**
+ * @covers \Trivago\Rumi\Models\CacheConfig
+ */
+class CacheConfigTest extends \PHPUnit_Framework_TestCase
 {
-    const CONFIG = 'config';
-    const CONFIG_SHORT = 'c';
 
-    const DEFAULT_CONFIG = '.rumi.yml';
+    public function testGivenValidConfig_WhenEntityCreated_ThenGetDirectoriesReturnsContent()
+    {
+        //given
+        $directories =  ['a', 'b', 'c'];
 
-    protected function configure(){
-        $this->addOption(
-            self::CONFIG,
-            self::CONFIG_SHORT,
-            InputOption::VALUE_REQUIRED,
-            'Configuration file to read',
-            self::DEFAULT_CONFIG);
+        // when
+        $cacheConfig = new CacheConfig($directories);
+
+        // then
+        $this->assertEquals($directories, iterator_to_array($cacheConfig, true));
+        $this->assertEquals(3, $cacheConfig->count());
     }
+
 }